OMA-51 Update educandu to v0.27.0
