+++ src/bundles/bundle-config.js
@@ -1,13 +1,13 @@
 import SiteLogo from './site-logo.js';
 import PageTemplate from './page-template.js';
-import HomePageLogo from './home-page-logo.js';
+import HomePageTemplate from './home-page-template.js';
 
 export default {
   getPageTemplateComponent() {
     return PageTemplate;
   },
-  getHomePageLogoComponent() {
-    return HomePageLogo;
+  getHomePageTemplateComponent() {
+    return HomePageTemplate;
   },
   getSiteLogoComponent() {
     return SiteLogo;
