+++ src/bundles/home-page-illustration.js
@@ -1,9 +1,8 @@
 /* eslint-disable max-lines */
 import React from 'react';
-import SiteLogo from './site-logo.js';
 
-function HomePageLogo() {
-  const renderLogo = () => (
+function HomePageIllustration() {
+  return (
     <svg
       version="1.1"
       x="0px"
@@ -18,7 +17,7 @@ function HomePageLogo() {
       <rect
         x="0"
         y="40.613098"
-        className="HomePageLogo-svgSt0"
+        className="HomePageIllustration-svgSt0"
         width="260.29999"
         height="168.73"
         />
@@ -26,108 +25,108 @@ function HomePageLogo() {
         transform="translate(-78.910004,-278.18689)"
         >
         <path
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           d="M 259.25,336.37 H 92.43 c 0,0 -0.31,12.38 0,12.66 h 166.82"
           />
         <line
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           x1="92.419998"
           y1="340.59"
           x2="259.25"
           y2="340.59"
           />
         <line
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           x1="92.419998"
           y1="344.81"
           x2="259.25"
           y2="344.81"
           />
         <path
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           d="M 259.25,361.05 H 92.43 c 0,0 -0.31,12.38 0,12.66 h 166.82"
           />
         <line
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           x1="92.419998"
           y1="365.26999"
           x2="259.25"
           y2="365.26999"
           />
         <line
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           x1="92.419998"
           y1="369.48999"
           x2="259.25"
           y2="369.48999"
           />
         <path
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           d="M 259.25,385.73 H 92.43 c 0,0 -0.31,12.38 0,12.66 h 166.82"
           />
         <line
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           x1="92.419998"
           y1="389.95001"
           x2="259.25"
           y2="389.95001"
           />
         <line
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           x1="92.419998"
           y1="394.17001"
           x2="259.25"
           y2="394.17001"
           />
         <path
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           d="M 259.25,410.41 H 92.43 c 0,0 -0.31,12.38 0,12.66 h 166.82"
           />
         <line
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           x1="92.419998"
           y1="414.63"
           x2="259.25"
           y2="414.63"
           />
         <line
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           x1="92.419998"
           y1="418.85001"
           x2="259.25"
           y2="418.85001"
           />
         <path
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           d="M 259.25,435.09 H 92.43 c 0,0 -0.31,12.38 0,12.66 h 166.82"
           />
         <line
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           x1="92.419998"
           y1="439.31"
           x2="259.25"
           y2="439.31"
           />
         <line
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           x1="92.419998"
           y1="443.54001"
           x2="259.25"
           y2="443.54001"
           />
         <path
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           d="M 259.25,459.78 H 92.43 c 0,0 -0.31,12.38 0,12.66 h 166.82"
           />
         <line
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           x1="92.419998"
           y1="464"
           x2="259.25"
           y2="464"
           />
         <line
-          className="HomePageLogo-svgSt1"
+          className="HomePageIllustration-svgSt1"
           x1="92.419998"
           y1="468.22"
           x2="259.25"
@@ -135,190 +134,190 @@ function HomePageLogo() {
           />
       </g>
       <path
-        className="HomePageLogo-svgSt2"
+        className="HomePageIllustration-svgSt2"
         d="m 87.249996,159.82311 c 4.68,1.73 5.26,4.82 5.26,4.82 0,0 0.56,-0.04 1.49,-0.1 l 5.12,44.69 H 115.04 c -2.61,-20.67 -4.83,-38.25 -5.8,-45.96 0.33,-0.03 0.66,-0.07 0.97,-0.1 0.78,-7.32 -0.23,-11.44 -0.41,-11.62 l -1.03,-35.54 c -0.17,-2.2 1.36,-4.55 -0.78,-4.31 l -0.19,0.02 -0.31,-2.05 c -0.39,-2.54 -2.63,-4.35 -5.15,-4.15 l -13.290004,1.07 c -5.5,0.44 -9.51,5.53 -8.74,11.11 2.07,15.13 5.86,41.72 6.94,42.12 z"
         />
       <path
-        className="HomePageLogo-svgSt3"
+        className="HomePageIllustration-svgSt3"
         d="m 408.85,209.22311 16.59,-1.99 c 0,0 33.42,-31.51 12.42,-73.52 -21,-42.01 -65.88,-54.42 -98.35,-53.47 -15.28,-30.56 -38.2,-56.34 -76.38,-56.34 -38.2,0 -58.24,24.82 -68.75,37.24 l -27.96,83.59 c -7.73,23.11 4.41,48.17 27.33,56.44 l 22.31,8.05 z"
         />
       <rect
         x="44.365067"
         y="137.69391"
         transform="rotate(-13.98066)"
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         width="5.7201333"
         height="2.2600527"
         />
       <path
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         d="m 76.619996,119.94311 1.99,8.02 c 0.09,0.38 -0.14,0.77 -0.52,0.86 l -0.46,0.12 c -2.09,0.52 -4.2,-0.75 -4.71,-2.83 l -0.46,-1.85 c -0.52,-2.09 0.75,-4.2 2.83,-4.71 l 0.46,-0.12 c 0.39,-0.11 0.78,0.12 0.87,0.51 z"
         />
       <rect
         x="45.21619"
         y="164.61668"
         transform="rotate(-13.98066)"
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         width="5.7201333"
         height="2.2600527"
         />
       <path
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         d="m 84.039996,146.15311 1.73,6.95 c 0.09,0.38 -0.14,0.77 -0.52,0.86 l -0.46,0.12 c -2.09,0.52 -4.2,-0.75 -4.71,-2.83 l -0.2,-0.78 c -0.52,-2.09 0.75,-4.2 2.83,-4.71 l 0.46,-0.12 c 0.39,-0.11 0.78,0.13 0.87,0.51 z"
         />
       <rect
         x="44.331196"
         y="150.70924"
         transform="rotate(-13.98066)"
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         width="5.7201333"
         height="2.2600527"
         />
       <path
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         d="m 79.869996,133.18311 1.69,6.8 -1.85,0.46 c -1.7,0.43 -3.42,-0.61 -3.85,-2.31 l -0.16,-0.64 c -0.43,-1.7 0.61,-3.42 2.31,-3.85 z"
         />
       <rect
         x="43.030907"
         y="137.62863"
         transform="rotate(-13.98066)"
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         width="5.7201333"
         height="2.2600527"
         />
       <path
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         d="m 75.309996,120.20311 1.99,8.02 c 0.09,0.38 -0.14,0.77 -0.52,0.86 l -0.46,0.12 c -2.09,0.52 -4.2,-0.75 -4.71,-2.83 l -0.46,-1.85 c -0.52,-2.09 0.75,-4.2 2.83,-4.71 l 0.46,-0.12 c 0.39,-0.1 0.77,0.13 0.87,0.51 z"
         />
       <rect
         x="43.880413"
         y="164.56166"
         transform="rotate(-13.98066)"
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         width="5.7201333"
         height="2.2600527"
         />
       <path
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         d="m 82.719996,146.41311 1.73,6.95 c 0.09,0.38 -0.14,0.77 -0.52,0.86 l -0.46,0.12 c -2.09,0.52 -4.2,-0.75 -4.71,-2.83 l -0.2,-0.78 c -0.52,-2.09 0.75,-4.2 2.83,-4.71 l 0.46,-0.12 c 0.39,-0.11 0.78,0.13 0.87,0.51 z"
         />
       <rect
         x="42.976765"
         y="150.64258"
         transform="rotate(-13.98066)"
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         width="5.7201333"
         height="2.2600527"
         />
       <path
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         d="m 78.559996,133.44311 1.69,6.8 -1.85,0.46 c -1.7,0.43 -3.42,-0.61 -3.85,-2.31 l -0.16,-0.64 c -0.43,-1.7 0.61,-3.42 2.31,-3.85 z"
         />
       <rect
         x="85.701859"
         y="160.03769"
         transform="rotate(-11.420052)"
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         width="5.7199883"
         height="2.2599955"
         />
       <path
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         d="m 121.46,144.04311 -1.64,-8.1 c -0.08,-0.39 0.17,-0.77 0.56,-0.84 l 0.47,-0.09 c 2.1,-0.43 4.16,0.94 4.58,3.04 l 0.38,1.87 c 0.43,2.1 -0.94,4.16 -3.04,4.58 l -0.47,0.09 c -0.38,0.08 -0.76,-0.17 -0.84,-0.55 z"
         />
       <rect
         x="84.846115"
         y="133.11658"
         transform="rotate(-11.420052)"
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         width="5.7199883"
         height="2.2599955"
         />
       <path
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         d="m 115.22,117.52311 -1.42,-7.02 c -0.08,-0.39 0.17,-0.77 0.56,-0.84 l 0.47,-0.09 c 2.1,-0.43 4.16,0.94 4.58,3.04 l 0.16,0.79 c 0.43,2.1 -0.94,4.16 -3.04,4.58 l -0.47,0.09 c -0.38,0.09 -0.76,-0.16 -0.84,-0.55 z"
         />
       <rect
         x="85.74794"
         y="147.02934"
         transform="rotate(-11.420052)"
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         width="5.7199883"
         height="2.2599955"
         />
       <path
-        className="HomePageLogo-svgSt4"
+        className="HomePageIllustration-svgSt4"
         d="m 118.81,130.67311 -1.39,-6.87 1.87,-0.38 c 1.72,-0.35 3.39,0.77 3.74,2.48 l 0.13,0.65 c 0.35,1.72 -0.77,3.39 -2.48,3.74 z"
         />
       <path
-        className="HomePageLogo-svgSt5"
+        className="HomePageIllustration-svgSt5"
         d="m 90.649996,156.70311 c -1.12,-0.27 -7.97,-25.68 -11.8,-40.18 -1.41,-5.33 1.99,-10.74 7.4,-11.79 l 13.07,-2.55 c 2.470004,-0.48 4.910004,1.02 5.590004,3.45 l 0.55,1.96 0.19,-0.04 c 2.1,-0.48 4.18,0.88 4.6,2.99 l 7.79,39.19 c 0.2,0.15 -4.28,0.4 -4.2,7.61 -6.31,1.4 -17.400004,3.44 -17.400004,3.44 0,0 -0.95,-2.93 -5.79,-4.08 z"
         />
       <path
-        className="HomePageLogo-svgSt6"
+        className="HomePageIllustration-svgSt6"
         d="m 243.46,60.46311 c 10.31,9.08 16.08,40.05 16.08,67.94 0,27.89 5.47,39.88 6.26,41.18 0.78,1.3 -65.43,4.95 -87.84,4.17 -22.42,-0.78 -32.85,-3.13 -32.85,-3.13 0,0 -0.78,-82.37 14.6,-103.22"
         />
       <path
-        className="HomePageLogo-svgSt7"
+        className="HomePageIllustration-svgSt7"
         d="m 230.74,182.51311 -1.35,-31.82 c 2.77,-4.99 7.78,-7.52 9.06,-25.06 0,0 10.58,-45.74 0.69,-61.19 -3.3,-5.15 -20.05,-18.41 -33.36,-18.28 -44.79,0.46 -47.43,43.5 -47.43,43.5 0,0 -5.18,52.55 7.95,63.94 10.9,9.46 19.57,11.48 19.57,11.48 l -1.05,18.43"
         />
       <path
-        className="HomePageLogo-svgSt8"
+        className="HomePageIllustration-svgSt8"
         d="m 190.74,102.72311 c 0.07,5.22 -4.5,15.52 -4.8,17.31 -0.17,0.99 0.57,2.76 2.01,3.94 1.44,1.18 3.28,1.84 3.28,1.84"
         />
       <path
-        className="HomePageLogo-svgSt5"
+        className="HomePageIllustration-svgSt5"
         d="m 125.91,209.22311 c -7.61,-30.17 -14.04,-55.68 -14.04,-55.68 l -14.420004,5.1 12.120004,50.59 z"
         />
       <path
-        className="HomePageLogo-svgSt9"
+        className="HomePageIllustration-svgSt9"
         d="m 204.43,106.34311 c 0,0 3.51,4.1 10.32,1.49"
         />
       <path
-        className="HomePageLogo-svgSt10"
+        className="HomePageIllustration-svgSt10"
         d="m 181.88,136.28311 c 0,0 8.02,6.66 16.48,1.33"
         />
       <path
-        className="HomePageLogo-svgSt10"
+        className="HomePageIllustration-svgSt10"
         d="M 158.32,90.53311"
         />
       <path
-        className="HomePageLogo-svgSt11"
+        className="HomePageIllustration-svgSt11"
         d="m 236.53,106.13311 c 2.04,-2.85 8.04,-7.2 10.41,-2.06 2.37,5.14 -0.71,17.88 -4.15,19.46 -3.43,1.57 -6.62,1.6 -6.62,1.6"
         />
       <path
-        className="HomePageLogo-svgSt10"
+        className="HomePageIllustration-svgSt10"
         d="m 240.47,108.08311 c 0,0 3.12,1.01 2.75,4.41 -0.28,2.6 -0.28,2.6 -0.28,2.6"
         />
       <path
-        className="HomePageLogo-svgSt10"
+        className="HomePageIllustration-svgSt10"
         d="m 239.49,112.56311 c 0,0 -1.19,1.67 -0.91,2.37"
         />
       <ellipse
-        className="HomePageLogo-svgSt6"
+        className="HomePageIllustration-svgSt6"
         cx="173.64"
         cy="105.6431"
         rx="2.8499999"
         ry="3.6700001"
         />
       <path
-        className="HomePageLogo-svgSt6"
+        className="HomePageIllustration-svgSt6"
         d="m 156.78,100.17311 c 0,0 23.32,-17.49 31.56,-33.97 7.03,8.84 24.92,20.1 37.79,19.7 1.81,7.44 7.24,16.08 11.06,18.49 4.22,-2.21 7.03,-2.21 7.03,-2.21 l 4.02,4.83 c 0,0 8.24,-35.38 -7.24,-49.04 -15.48,-13.67 -36.38,-19.09 -54.07,-13.46 -17.69,5.63 -38.39,22.49 -30.15,55.66 z"
         />
       <path
-        className="HomePageLogo-svgSt12"
+        className="HomePageIllustration-svgSt12"
         d="m 228.67,113.31311 c 0,0.28 -0.38,11.24 -0.38,11.24 l 7.18,-0.28 c 0,0 2.74,-0.85 2.83,-1.23 0.09,-0.38 -0.19,-4.15 -0.19,-4.15 z"
         />
       <path
-        className="HomePageLogo-svgSt9"
+        className="HomePageIllustration-svgSt9"
         d="m 206.42,94.43311 c 0,0 7.13,-0.35 11.47,3.65"
         />
       <path
-        className="HomePageLogo-svgSt9"
+        className="HomePageIllustration-svgSt9"
         d="m 176.71,92.18311 c 0,0 -7.13,-0.35 -11.47,3.65"
         />
       <path
-        className="HomePageLogo-svgSt6"
+        className="HomePageIllustration-svgSt6"
         d="m 185.36,163.86311 c 0,0 11.19,1.97 14.83,1.81 -11.05,2.42 -11.35,3.78 -21.18,0.91"
         />
       <g
@@ -326,60 +325,60 @@ function HomePageLogo() {
         >
         <g>
           <path
-            className="HomePageLogo-svgSt6"
+            className="HomePageIllustration-svgSt6"
             d="m 219.86,402.29 c 3.63,3.93 8.32,6.68 13.56,7.66 -1.37,-12.66 -0.94,-26.93 -0.44,-35.69 4.28,-3.47 11.18,-9.36 17.64,-16.23 -3.77,-2.43 -8.16,-3.82 -12.84,-3.82 -3.87,0 -7.53,0.95 -10.81,2.65 -0.02,0.19 -0.04,0.38 -0.06,0.56 -3.58,12.53 -5.74,29.05 -7.05,44.87 z"
             />
           <path
-            className="HomePageLogo-svgSt12"
+            className="HomePageIllustration-svgSt12"
             d="m 232.98,374.27 c -0.5,8.76 -0.93,23.03 0.44,35.69 1.41,0.27 2.87,0.41 4.35,0.41 14.09,0 25.51,-12.57 25.51,-28.08 0,-10.35 -5.1,-19.39 -12.67,-24.26 -6.45,6.88 -13.34,12.77 -17.63,16.24 z m 15.86,15.93 c -1.71,0 -3.1,-2.45 -3.1,-5.47 0,-3.02 1.39,-5.47 3.1,-5.47 1.71,0 3.1,2.45 3.1,5.47 0,3.02 -1.39,5.47 -3.1,5.47 z m 3.38,-22.55 c 0.54,0.02 0.96,0.49 0.93,1.02 -0.02,0.54 -0.5,0.98 -1.02,0.93 -0.07,0 -7.39,-0.3 -11.74,3.7 -0.19,0.17 -0.43,0.26 -0.67,0.26 -0.27,0 -0.53,-0.1 -0.72,-0.31 -0.36,-0.4 -0.34,-1.02 0.06,-1.39 4.96,-4.55 12.83,-4.23 13.16,-4.21 z"
             />
           <ellipse
-            className="HomePageLogo-svgSt6"
+            className="HomePageIllustration-svgSt6"
             cx="248.84"
             cy="384.73001"
             rx="3.0999999"
             ry="5.4699998"
             />
           <path
-            className="HomePageLogo-svgSt6"
+            className="HomePageIllustration-svgSt6"
             d="m 239.01,373.25 c 0.19,0.21 0.46,0.31 0.72,0.31 0.23,0 0.48,-0.09 0.67,-0.26 4.35,-4 11.66,-3.7 11.74,-3.7 0.52,0.04 1,-0.39 1.02,-0.93 0.02,-0.54 -0.39,-1 -0.93,-1.02 -0.33,-0.02 -8.2,-0.35 -13.16,4.22 -0.41,0.37 -0.44,0.98 -0.06,1.38 z"
             />
         </g>
       </g>
       <ellipse
-        className="HomePageLogo-svgSt13"
+        className="HomePageIllustration-svgSt13"
         cx="209.66"
         cy="125.8031"
         rx="2.9100001"
         ry="3.27"
         />
       <ellipse
-        className="HomePageLogo-svgSt13"
+        className="HomePageIllustration-svgSt13"
         cx="175.06999"
         cy="125.5531"
         rx="2.9100001"
         ry="3.27"
         />
       <path
-        className="HomePageLogo-svgSt14"
+        className="HomePageIllustration-svgSt14"
         d="m 158.86,76.03311 c -14.09,0 -25.51,12.57 -25.51,28.08 0,10.78 5.52,20.14 13.63,24.84 l -10.85,41.08 6.7,1.77 10.63,-40.25 c 1.75,0.41 3.55,0.64 5.41,0.64 14.09,0 25.51,-12.57 25.51,-28.08 0,-15.51 -11.43,-28.08 -25.52,-28.08 z m 0,49.98 c -10.67,0 -19.34,-9.83 -19.34,-21.91 0,-12.08 8.68,-21.91 19.34,-21.91 10.66,0 19.34,9.83 19.34,21.91 0.01,12.09 -8.67,21.91 -19.34,21.91 z"
         />
       <ellipse
         transform="rotate(-16.988997)"
-        className="HomePageLogo-svgSt3"
+        className="HomePageIllustration-svgSt3"
         cx="303.28754"
         cy="130.69189"
         rx="6.1702523"
         ry="4.700192"
         />
       <polyline
-        className="HomePageLogo-svgSt15"
+        className="HomePageIllustration-svgSt15"
         points="411.84,313.91 411.84,292.98 416.02,292.98 "
         transform="translate(-78.910004,-278.18689)"
         />
       <ellipse
         transform="rotate(-16.988997)"
-        className="HomePageLogo-svgSt3"
+        className="HomePageIllustration-svgSt3"
         cx="328.38211"
         cy="131.2795"
         rx="6.1702523"
@@ -387,19 +386,19 @@ function HomePageLogo() {
         />
       <ellipse
         transform="rotate(-16.988997)"
-        className="HomePageLogo-svgSt3"
+        className="HomePageIllustration-svgSt3"
         cx="346.65454"
         cy="130.4088"
         rx="6.1702523"
         ry="4.700192"
         />
       <polyline
-        className="HomePageLogo-svgSt15"
+        className="HomePageIllustration-svgSt15"
         points="453.15,300.87 453.15,279.94 435.93,286.11 435.93,307.04 "
         transform="translate(-78.910004,-278.18689)"
         />
       <line
-        className="HomePageLogo-svgSt15"
+        className="HomePageIllustration-svgSt15"
         x1="373.57001"
         y1="5.5830989"
         x2="356.64999"
@@ -412,7 +411,7 @@ function HomePageLogo() {
 
           <ellipse
             transform="matrix(0.9564,-0.2922,0.2922,0.9564,-109.7963,153.1222)"
-            className="HomePageLogo-svgSt6"
+            className="HomePageIllustration-svgSt6"
             cx="457.72"
             cy="444.13"
             rx="6.1700001"
@@ -421,24 +420,24 @@ function HomePageLogo() {
         </g>
         <g>
           <polygon
-            className="HomePageLogo-svgSt6"
+            className="HomePageIllustration-svgSt6"
             points="461.12,421.48 466.54,421.48 466.54,423.95 463.59,423.95 463.59,443.65 461.12,443.65 "
             />
         </g>
       </g>
       <path
-        className="HomePageLogo-svgSt6"
+        className="HomePageIllustration-svgSt6"
         d="m 78.729996,197.04311 c 8.99,3.7 16.52,7.87 21.18,12.18 H 109.62 c 0.02,-0.1 0.04,-0.15 0.04,-0.15 0.33,-5.1 -10.950004,-36.44 -27.430004,-38.78 -16.5,-2.35 -17.41,21.03 -3.5,26.75 z"
         />
       <g
         transform="translate(-78.910004,-278.18689)"
         >
         <path
-          className="HomePageLogo-svgSt12"
+          className="HomePageIllustration-svgSt12"
           d="m 217.87,460.41 0.16,-1.35 0.07,-0.14 c 7.95,-16.02 3.35,-21.89 1.83,-23.82 l -0.72,-0.93 0.89,-0.54 c 1.15,-0.7 3,0 5.68,1.15 0.56,0.24 1.09,0.47 1.51,0.62 0.24,0.09 0.44,0.06 0.68,-0.09 0.73,-0.45 1.36,-1.81 1.47,-2.76 0.07,-0.59 0.17,-3.55 -4.55,-4.54 -4.25,-0.89 -7.45,1.15 -7.49,1.17 l -2.55,1.67 1.2,-2.8 c 0.07,-0.17 1.86,-4.25 6,-5.4 1.16,-0.32 1.83,-0.86 2.05,-1.65 0.35,-1.24 -0.49,-2.8 -0.8,-3.13 -0.09,-0.1 -0.6,-0.35 -1.76,-0.3 -2.1,0.08 -5,1.1 -6.46,2.9 -2.57,3.15 -3.78,8.07 -4.28,10.72 l -1.07,5.77 -0.75,-5.82 c -0.52,-3.99 0.13,-11 1.04,-14.36 1.14,-4.25 1.14,-7.02 0.77,-7.55 -0.19,-0.28 -0.77,-0.73 -1.6,-0.86 -0.7,-0.1 -1.36,0.06 -1.95,0.49 -1.33,1.88 -3.11,6.51 -4.22,17.83 -0.49,4.99 1.73,16.6 2.97,22.04 0.81,3.55 1.11,14.7 0.83,17.21 0.65,7.16 1.6,14.46 2.75,21.48 h 19.81 C 224.87,473.31 219.2,462.5 218.04,460.7 Z"
           />
         <path
-          className="HomePageLogo-svgSt6"
+          className="HomePageIllustration-svgSt6"
           d="m 209.58,487.41 c -1.15,-7.02 -2.09,-14.32 -2.75,-21.48 0.27,-2.51 -0.02,-13.66 -0.83,-17.21 -1.25,-5.44 -3.47,-17.06 -2.97,-22.04 1.11,-11.32 2.89,-15.94 4.22,-17.83 0.59,-0.43 1.25,-0.6 1.95,-0.49 0.83,0.12 1.41,0.58 1.6,0.86 0.36,0.53 0.37,3.3 -0.77,7.55 -0.9,3.36 -1.55,10.37 -1.04,14.36 l 0.75,5.82 1.07,-5.77 c 0.49,-2.65 1.7,-7.57 4.28,-10.72 1.46,-1.8 4.36,-2.81 6.46,-2.9 1.16,-0.04 1.67,0.2 1.76,0.3 0.31,0.33 1.14,1.89 0.8,3.13 -0.22,0.78 -0.89,1.33 -2.05,1.65 -4.13,1.14 -5.92,5.22 -6,5.4 l -1.2,2.8 2.55,-1.67 c 0.03,-0.02 3.24,-2.06 7.49,-1.17 4.72,0.99 4.62,3.96 4.55,4.54 -0.11,0.94 -0.75,2.31 -1.47,2.76 -0.24,0.15 -0.44,0.17 -0.68,0.09 -0.42,-0.15 -0.95,-0.38 -1.51,-0.62 -2.68,-1.15 -4.53,-1.84 -5.68,-1.15 l -0.89,0.54 0.72,0.93 c 1.51,1.93 6.12,7.8 -1.83,23.82 l -0.07,0.14 -0.16,1.35 0.18,0.28 c 1.17,1.81 6.83,12.61 11.34,26.72 h 1.93 c -4.42,-13.99 -10.01,-24.9 -11.54,-27.44 l 0.04,-0.38 c 6.95,-14.07 4.6,-20.96 2.64,-24.13 0.88,0.29 1.94,0.74 2.59,1.02 0.59,0.25 1.14,0.49 1.6,0.65 0.77,0.28 1.56,0.19 2.28,-0.25 1.3,-0.8 2.18,-2.72 2.34,-4.12 0.28,-2.39 -1.12,-5.55 -6,-6.57 -2.3,-0.48 -4.31,-0.24 -5.85,0.17 0.77,-0.8 1.8,-1.57 3.11,-1.93 2.26,-0.63 3.06,-1.94 3.34,-2.94 0.59,-2.13 -0.7,-4.34 -1.25,-4.9 -0.76,-0.8 -2.14,-0.91 -3.17,-0.88 -2.52,0.1 -5.92,1.24 -7.82,3.58 -1.22,1.49 -2.15,3.3 -2.88,5.14 0.19,-2.75 0.58,-5.5 1.03,-7.18 0.44,-1.65 1.8,-7.2 0.51,-9.07 -0.56,-0.82 -1.68,-1.46 -2.84,-1.64 -0.87,-0.13 -2.18,-0.07 -3.47,0.94 l -0.1,0.08 -0.07,0.1 c -2.14,2.91 -3.7,9.25 -4.65,18.84 -0.64,6.53 2.86,21.98 3.02,22.63 0.77,3.38 1.07,14.53 0.79,16.67 l -0.01,0.1 0.01,0.1 c 0.66,7.19 1.59,14.45 2.71,21.41 z"
           />
       </g>
@@ -446,27 +445,27 @@ function HomePageLogo() {
         transform="translate(-78.910004,-278.18689)"
         >
         <polygon
-          className="HomePageLogo-svgSt16"
+          className="HomePageIllustration-svgSt16"
           points="171.11,427.69 178.5,438.59 190.08,487.41 190.47,487.41 178.86,438.47 171.42,427.47 "
           />
         <polygon
-          className="HomePageLogo-svgSt16"
+          className="HomePageIllustration-svgSt16"
           points="166.57,413.63 180.78,437.96 192.51,487.41 192.9,487.41 181.14,437.84 166.9,413.44 "
           />
         <polygon
-          className="HomePageLogo-svgSt16"
+          className="HomePageIllustration-svgSt16"
           points="162.73,400.75 183.06,437.32 194.94,487.41 195.33,487.41 183.42,437.21 163.06,400.56 "
           />
         <polygon
-          className="HomePageLogo-svgSt16"
+          className="HomePageIllustration-svgSt16"
           points="185.07,393.43 185.33,436.67 197.37,487.41 197.76,487.41 185.71,436.62 185.45,393.42 "
           />
         <path
-          className="HomePageLogo-svgSt16"
+          className="HomePageIllustration-svgSt16"
           d="m 187.99,435.96 c -0.02,-0.67 0.25,-18.83 0.41,-28.4 l -0.38,-0.01 c -0.38,24.1 -0.44,27.85 -0.36,28.48 l -0.04,0.01 12.19,51.37 h 0.39 z"
           />
         <path
-          className="HomePageLogo-svgSt16"
+          className="HomePageIllustration-svgSt16"
           d="m 190.27,435.33 -0.02,0.01 c 0,-0.01 0,-0.01 0.01,-0.02 l 0.01,-0.01 -0.01,0.01 c 0.12,-0.7 0.25,-3.45 0.7,-14.36 l -0.38,-0.02 c -0.16,3.87 -0.59,14 -0.67,14.4 l 0.15,0.04 -0.16,0.05 12.33,52 h 0.39 z"
           />
       </g>
@@ -474,67 +473,67 @@ function HomePageLogo() {
         transform="translate(-78.910004,-278.18689)"
         >
         <circle
-          className="HomePageLogo-svgSt4"
+          className="HomePageIllustration-svgSt4"
           cx="165.8"
           cy="399.42999"
           r="3.3299999"
           />
         <circle
-          className="HomePageLogo-svgSt4"
+          className="HomePageIllustration-svgSt4"
           cx="169.66"
           cy="412.23001"
           r="3.3299999"
           />
         <circle
-          className="HomePageLogo-svgSt4"
+          className="HomePageIllustration-svgSt4"
           cx="173.55"
           cy="425.28"
           r="3.3299999"
           />
         <circle
-          className="HomePageLogo-svgSt4"
+          className="HomePageIllustration-svgSt4"
           cx="182.33"
           cy="394.54999"
           r="3.3299999"
           />
         <circle
-          className="HomePageLogo-svgSt4"
+          className="HomePageIllustration-svgSt4"
           cx="184.87"
           cy="407.67001"
           r="3.3299999"
           />
         <circle
-          className="HomePageLogo-svgSt4"
+          className="HomePageIllustration-svgSt4"
           cx="187.42999"
           cy="421.04999"
           r="3.3299999"
           />
       </g>
       <path
-        className="HomePageLogo-svgSt6"
+        className="HomePageIllustration-svgSt6"
         d="m 279.49,209.22311 c -8.91,-16.48 -30.42,-22.18 -49.46,-30.43 -20.64,11.3 -45.7,0.25 -45.7,0.25 -17.16,8.68 -29.08,19.25 -37.31,30.18 z"
         />
       <path
-        className="HomePageLogo-svgSt6"
+        className="HomePageIllustration-svgSt6"
         d="m 334.03,209.22311 h 10.18 c 5.89,-3.49 15.61,-5.67 27.4,-4.41 17.79,1.9 16.6,-15.18 10.67,-18.5 -5.93,-3.32 -16.84,-2.61 -27.51,2.84 -0.28,0.14 -0.56,0.3 -0.84,0.46 -0.11,-1.67 0.52,-3.96 3.96,-5.82 6.16,-3.32 10.2,-13.29 8.54,-19.92 -1.66,-6.64 -7.11,-8.3 -14.23,-6.64 -7.11,1.66 -7.55,11.75 -7.11,15.18 0.68,5.46 -0.23,12.57 -1.9,14.23 -1.66,1.66 -4.98,-0.23 -6.64,-4.5 -1.66,-4.27 -7.11,-2.84 -9.02,1.66 -1.9,4.5 4.03,9.96 6.4,15.65 1.12,2.67 0.78,6.48 0.1,9.77 z"
         />
       <path
-        className="HomePageLogo-svgSt17"
+        className="HomePageIllustration-svgSt17"
         d="m 295.87,191.03311 c 6.63,-2.58 8.47,-0.17 9.08,0.68 0.62,0.86 -0.17,3.38 -2.15,4.02 -1.98,0.64 -7.24,3.88 -5.55,6.66 1.7,2.79 4.94,2.84 6.83,0.97 1.89,-1.87 5.82,-2.13 7.89,-0.96 1.55,0.89 6.11,3.49 8.9,6.82 h 7.34 c -0.88,-5.79 -2.81,-17.28 -4.96,-21.87 -2.94,-6.24 -8.5,-17.15 -25.21,-16.41 -16.71,0.74 -16.48,13.12 -15.88,16.25 0.6,3.12 7.08,6.42 13.71,3.84 z"
         />
       <g
         transform="translate(-78.910004,-278.18689)"
         >
         <path
-          className="HomePageLogo-svgSt4"
+          className="HomePageIllustration-svgSt4"
           d="m 196.55,452.45 0.02,0.1 c 0.06,0.28 -0.11,0.56 -0.39,0.62 l -15.72,3.6 c -0.28,0.06 -0.56,-0.11 -0.62,-0.39 l -0.02,-0.1 c -0.06,-0.28 0.11,-0.56 0.39,-0.62 l 15.72,-3.6 c 0.28,-0.06 0.55,0.11 0.62,0.39 z"
           />
         <path
-          className="HomePageLogo-svgSt4"
+          className="HomePageIllustration-svgSt4"
           d="m 199.95,466.39 0.02,0.1 c 0.06,0.28 -0.11,0.56 -0.39,0.62 l -15.72,3.6 c -0.28,0.06 -0.56,-0.11 -0.62,-0.39 l -0.02,-0.1 c -0.06,-0.28 0.11,-0.56 0.39,-0.62 l 15.72,-3.6 c 0.28,-0.07 0.55,0.11 0.62,0.39 z"
           />
         <path
-          className="HomePageLogo-svgSt4"
+          className="HomePageIllustration-svgSt4"
           d="m 203.35,480.32 0.02,0.1 c 0.06,0.28 -0.11,0.56 -0.39,0.62 l -15.72,3.6 c -0.28,0.06 -0.56,-0.11 -0.62,-0.39 l -0.02,-0.1 c -0.06,-0.28 0.11,-0.56 0.39,-0.62 l 15.72,-3.6 c 0.28,-0.06 0.56,0.12 0.62,0.39 z"
           />
       </g>
@@ -542,27 +541,20 @@ function HomePageLogo() {
         transform="translate(-78.910004,-278.18689)"
         >
         <path
-          className="HomePageLogo-svgSt4"
+          className="HomePageIllustration-svgSt4"
           d="m 193.1,435.07 0.02,0.11 c 0.07,0.3 -0.12,0.6 -0.43,0.67 l -16.59,3.8 c -0.3,0.07 -0.6,-0.12 -0.67,-0.43 l -0.02,-0.11 c -0.07,-0.3 0.12,-0.6 0.43,-0.67 l 16.59,-3.8 c 0.3,-0.06 0.61,0.13 0.67,0.43 z"
           />
       </g>
       <path
-        className="HomePageLogo-svgSt6"
+        className="HomePageIllustration-svgSt6"
         d="m 78.729996,197.04311 c 8.99,3.7 16.52,7.87 21.18,12.18 H 109.62 c 0.02,-0.1 0.04,-0.15 0.04,-0.15 0.33,-5.1 -10.950004,-36.44 -27.430004,-38.78 -16.5,-2.35 -17.41,21.03 -3.5,26.75 z"
         />
       <path
-        className="HomePageLogo-svgSt18"
+        className="HomePageIllustration-svgSt18"
         d="m 106.31,209.17311 c -8.190004,-20.66 -26.780004,-32.44 -33.850004,-36.2"
         />
     </svg>
   );
-
-  return (
-    <div className="HomePageLogo">
-      {renderLogo()}
-      <SiteLogo />
-    </div>
-  );
 }
 
-export default HomePageLogo;
+export default HomePageIllustration;
