+++ src/bundles/home-page-illustration.less
@@ -1,0 +1,19 @@
+.HomePageIllustration-svgSt0 {fill:#F7F8FB;}
+.HomePageIllustration-svgSt1 {fill:none;stroke:#CCD5E2;stroke-width:0.6359;stroke-miterlimit:10;}
+.HomePageIllustration-svgSt2 {opacity:0.2;fill:#302D7E;enable-background:new    ;}
+.HomePageIllustration-svgSt3 {fill:#D9E2F4;}
+.HomePageIllustration-svgSt4 {fill:#EF825E;}
+.HomePageIllustration-svgSt5 {fill:#F0B7A8;}
+.HomePageIllustration-svgSt6 {fill:#302D7E;}
+.HomePageIllustration-svgSt7 {fill:#FEEDDD;stroke:#302D7E;stroke-width:2.47;stroke-linecap:square;stroke-linejoin:round;stroke-miterlimit:10;}
+.HomePageIllustration-svgSt8 {fill:none;stroke:#302D7E;stroke-width:2.1565;stroke-linecap:round;stroke-miterlimit:10;}
+.HomePageIllustration-svgSt9 {fill:none;stroke:#302D7E;stroke-width:1.8083;stroke-linecap:round;stroke-miterlimit:10;}
+.HomePageIllustration-svgSt10 {fill:none;stroke:#302D7E;stroke-width:2.47;stroke-linecap:round;stroke-miterlimit:10;}
+.HomePageIllustration-svgSt11 {fill:#FEEDDD;stroke:#302D7E;stroke-width:2.47;stroke-linecap:round;stroke-linejoin:round;stroke-miterlimit:10;}
+.HomePageIllustration-svgSt12 {fill:#FEEDDD;}
+.HomePageIllustration-svgSt13 {opacity:0.23;fill:#EF825E;enable-background:new    ;}
+.HomePageIllustration-svgSt14 {fill:#EF825E;stroke:#302D7E;stroke-width:2.4682;stroke-linecap:round;stroke-miterlimit:10;}
+.HomePageIllustration-svgSt15 {fill:none;stroke:#D9E2F4;stroke-width:2.4682;stroke-miterlimit:10;}
+.HomePageIllustration-svgSt16 {fill:#FCE2DD;}
+.HomePageIllustration-svgSt17 {fill:#302D7E;stroke:#302D7E;stroke-width:2.1732;stroke-linecap:round;stroke-linejoin:round;stroke-miterlimit:10;}
+.HomePageIllustration-svgSt18 {fill:none;stroke:#D9E2F4;stroke-width:0.617;stroke-miterlimit:10;}
