--- src/bundles/home-page-logo.less
@@ -1,25 +1,0 @@
-.HomePageLogo {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-}
-
-.HomePageLogo-svgSt0 {fill:#F7F8FB;}
-.HomePageLogo-svgSt1 {fill:none;stroke:#CCD5E2;stroke-width:0.6359;stroke-miterlimit:10;}
-.HomePageLogo-svgSt2 {opacity:0.2;fill:#302D7E;enable-background:new    ;}
-.HomePageLogo-svgSt3 {fill:#D9E2F4;}
-.HomePageLogo-svgSt4 {fill:#EF825E;}
-.HomePageLogo-svgSt5 {fill:#F0B7A8;}
-.HomePageLogo-svgSt6 {fill:#302D7E;}
-.HomePageLogo-svgSt7 {fill:#FEEDDD;stroke:#302D7E;stroke-width:2.47;stroke-linecap:square;stroke-linejoin:round;stroke-miterlimit:10;}
-.HomePageLogo-svgSt8 {fill:none;stroke:#302D7E;stroke-width:2.1565;stroke-linecap:round;stroke-miterlimit:10;}
-.HomePageLogo-svgSt9 {fill:none;stroke:#302D7E;stroke-width:1.8083;stroke-linecap:round;stroke-miterlimit:10;}
-.HomePageLogo-svgSt10 {fill:none;stroke:#302D7E;stroke-width:2.47;stroke-linecap:round;stroke-miterlimit:10;}
-.HomePageLogo-svgSt11 {fill:#FEEDDD;stroke:#302D7E;stroke-width:2.47;stroke-linecap:round;stroke-linejoin:round;stroke-miterlimit:10;}
-.HomePageLogo-svgSt12 {fill:#FEEDDD;}
-.HomePageLogo-svgSt13 {opacity:0.23;fill:#EF825E;enable-background:new    ;}
-.HomePageLogo-svgSt14 {fill:#EF825E;stroke:#302D7E;stroke-width:2.4682;stroke-linecap:round;stroke-miterlimit:10;}
-.HomePageLogo-svgSt15 {fill:none;stroke:#D9E2F4;stroke-width:2.4682;stroke-miterlimit:10;}
-.HomePageLogo-svgSt16 {fill:#FCE2DD;}
-.HomePageLogo-svgSt17 {fill:#302D7E;stroke:#302D7E;stroke-width:2.1732;stroke-linecap:round;stroke-linejoin:round;stroke-miterlimit:10;}
-.HomePageLogo-svgSt18 {fill:none;stroke:#D9E2F4;stroke-width:0.617;stroke-miterlimit:10;}
