+++ src/bundles/home-page-template.less
@@ -1,0 +1,10 @@
+.HomePageTemplate-logo {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  margin-bottom: 10px;
+}
+
+.HomePageTemplate-info {
+  .m-content;
+}
