+++ src/bundles/page-footer.less
@@ -1,0 +1,17 @@
+.PageFooter {
+  flex: none;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: @edu-content-padding;
+  position: relative;
+  z-index: @edu-z-index-base-layer;
+}
+
+.PageFooter-link {
+  display: inline-block;
+
+  &:not(:last-child)::after {
+    content: "\00A0\00A0|\00A0\00A0";
+  }
+}
