+++ src/bundles/page-header.less
@@ -1,0 +1,55 @@
+@page-header-content-min-height: 32px;
+@page-header-min-height: @page-header-content-min-height + (2* @edu-content-padding);
+
+.PageHeader {
+  top: 0;
+  left: 0;
+  width: 100%;
+  position: sticky;
+  z-index: @edu-z-index-cover-layer;
+  box-shadow: 0 2px 8px @grey-4;
+  margin-bottom: @edu-content-padding;
+
+  &.PageHeader--fullScreen {
+    box-shadow: none;
+  }
+}
+
+.PageHeader-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  background-color: @edu-white;
+  min-height: @page-header-min-height;
+}
+
+.PageHeader-headerContent {
+  display: block;
+  padding: @edu-content-padding;
+
+  &.PageHeader-headerContent--left {
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+  }
+
+  &.PageHeader-headerContent--right {
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+  }
+}
+
+.PageHeader-languageSwitch {
+  white-space: nowrap;
+}
+
+.PageHeader-headerButton {
+  .m-icon-button;
+  .m-grouped-button;
+}
+
+.PageHeader-loginButton {
+  cursor: pointer;
+  margin-right: (2.5 * @edu-button-spacing);
+}
