+++ src/main.less
@@ -3,8 +3,11 @@
 // Styles for OMA:
 @import url('./bundles/site-logo.less');
 @import url('./bundles/lato-font.less');
+@import url('./bundles/page-header.less');
+@import url('./bundles/page-footer.less');
 @import url('./bundles/page-template.less');
-@import url('./bundles/home-page-logo.less');
+@import url('./bundles/home-page-template.less');
+@import url('./bundles/home-page-illustration.less');
 
 @oma-blue: #002F65;
